+++ examples/wireless/wifi-eht-network.cc
@@ -138,10 +138,11 @@ main(int argc, char* argv[])
     bool useRts{false};
     bool use80Plus80{false};
     uint16_t mpduBufferSize{512};
+    std::string emlsrMgrTypeId{"ns3::DefaultEmlsrManager"};
     std::string emlsrLinks;
     uint16_t paddingDelayUsec{32};
     uint16_t transitionDelayUsec{128};
-    uint16_t channelSwitchDelayUsec{100};
+    Time channelSwitchDelay{"100us"};
     bool switchAuxPhy{true};
     uint16_t auxPhyChWidth{20};
     bool auxPhyTxCapable{true};
@@ -182,6 +183,7 @@ main(int argc, char* argv[])
         "Whether the third link operates in the 2.4, 5 or 6 GHz band (0 means the device has up to "
         "two links, otherwise the band must be different than first link and second link)",
         frequency3);
+    cmd.AddValue("emlsrMgrTypeId", "The ns-3 TypeId of the EMLSR manager to use", emlsrMgrTypeId);
     cmd.AddValue("emlsrLinks",
                  "The comma separated list of IDs of EMLSR links (for MLDs only)",
                  emlsrLinks);
@@ -201,9 +203,7 @@ main(int argc, char* argv[])
     cmd.AddValue("emlsrAuxTxCapable",
                  "Whether Aux PHYs are capable of transmitting.",
                  auxPhyTxCapable);
-    cmd.AddValue("channelSwitchDelay",
-                 "The PHY channel switch delay in microseconds",
-                 channelSwitchDelayUsec);
+    cmd.AddValue("channelSwitchDelay", "The PHY channel switch delay", channelSwitchDelay);
     cmd.AddValue("distance",
                  "Distance in meters between the station and the access point",
                  distance);
@@ -432,10 +432,10 @@ main(int argc, char* argv[])
 
                 SpectrumWifiPhyHelper phy(nLinks);
                 phy.SetPcapDataLinkType(WifiPhyHelper::DLT_IEEE802_11_RADIO);
-                phy.Set("ChannelSwitchDelay", TimeValue(MicroSeconds(channelSwitchDelayUsec)));
+                phy.Set("ChannelSwitchDelay", TimeValue(channelSwitchDelay));
 
                 mac.SetType("ns3::StaWifiMac", "Ssid", SsidValue(ssid));
-                mac.SetEmlsrManager("ns3::DefaultEmlsrManager",
+                mac.SetEmlsrManager(emlsrMgrTypeId,
                                     "EmlsrLinkSet",
                                     StringValue(emlsrLinks),
                                     "EmlsrPaddingDelay",